+++ src/entity/Admin.java
@@ -1,4 +1,4 @@
-package entities;
+package entity;
 public class Admin extends User{
 
     public boolean isCool = true;
